added comment posting and showing of basic comment data

+++ app/controllers/user_workout_comments_controller.rb
@@ -7,7 +7,7 @@ class UserWorkoutCommentsController < ApplicationController
   def create
     comment = UserWorkoutComment.new(
       workout_id: params[:workout_id],
-      user_id: params[:user_id],
+      user_id: current_user.id,
       comment: params[:comment],
     )
     comment.save
@@ -16,7 +16,9 @@ class UserWorkoutCommentsController < ApplicationController
 
   def destroy
     comment = UserWorkoutComment.find_by(id: params[:id])
-    comment.destroy
+    if current_user.id == comment.user_id
+      comment.destroy
+    end
     render json: { message: "Deleted Comment..." }
   end
 end
